The user does not need to know the existence of the Compressed struct
