+++ src/sparse_matrix/mod.rs
@@ -10,7 +10,7 @@
 pub mod compressed_matrix;
 pub mod triplet;
 
-pub use self::compressed_matrix::CompressedMatrix;
+pub use self::compressed_matrix::{CompressedMatrix, CompressedLinear, CompressedLinearMut};
 pub use self::compressed_matrix::csc_matrix::CscMatrix;
 pub use self::compressed_matrix::csr_matrix::CsrMatrix;
 pub use self::triplet::Triplet;
